+++ packages/zowe-explorer/__tests__/__unit__/trees/job/JobActions.unit.test.ts
@@ -1404,3 +1404,26 @@ describe("sortJobs function", () => {
         expect(jobsSortBy).not.toHaveBeenCalled();
     });
 });
+
+describe("copyName function", () => {
+    it("copies the job with format JobName(JobId)", async () => {
+        const node = new ZoweJobNode({ label: "JOBNAME(ID123456) - ACTIVE", collapsibleState: vscode.TreeItemCollapsibleState.Collapsed });
+        node.job = {
+            jobname: "JOBNAME",
+            jobid: "ID123456",
+        } as any;
+        node.contextValue = Constants.JOBS_JOB_CONTEXT;
+        const writeTextSpy = jest.spyOn(vscode.env.clipboard, "writeText");
+        await JobActions.copyName(node);
+        expect(writeTextSpy).toHaveBeenCalledWith("JOBNAME(ID123456)");
+        writeTextSpy.mockRestore();
+    });
+
+    it("copies a node's label for spools and nodes with missing job objects", async () => {
+        const node = new ZoweSpoolNode({ label: "JES2:JESMSGLG(2)", collapsibleState: vscode.TreeItemCollapsibleState.None });
+        const writeTextSpy = jest.spyOn(vscode.env.clipboard, "writeText");
+        await JobActions.copyName(node);
+        expect(writeTextSpy).toHaveBeenCalledWith("JES2:JESMSGLG(2)");
+        writeTextSpy.mockRestore();
+    });
+});