+++ packages/zowe-explorer/__tests__/__unit__/trees/shared/SharedHistoryView.unit.test.ts
@@ -250,7 +250,6 @@ describe("HistoryView Unit Tests", () => {
             const historyView = await initializeHistoryViewMock(blockMocks, globalMocks);
 
             expect(historyView["getHistoryData"]("ds")).toEqual({
-                dsTemplates: undefined,
                 favorites: undefined,
                 fileHistory: [],
                 search: undefined,