+++ packages/zowe-explorer/src/configuration/Profiles.ts
@@ -556,6 +556,12 @@ export class Profiles extends ProfilesCache {
 
         const returnValue: string[] = [promptInfo.profile.user, promptInfo.profile.password, promptInfo.profile.base64EncodedAuth];
         this.updateProfilesArrays(promptInfo);
+
+        // If secure credentials are enabled, the config file won't change after updating existing credentials
+        // (as the "secure" fields are already set). Fire the event emitter to notify extenders of the change.
+        if (SettingsConfig.getDirectValue<boolean>(Constants.SETTINGS_SECURE_CREDENTIALS_ENABLED)) {
+            ZoweExplorerApiRegister.getInstance().onProfilesUpdateEmitter.fire(Validation.EventType.UPDATE);
+        }
         return returnValue;
     }