+++ packages/zowe-explorer/src/trees/dataset/DatasetActions.ts
@@ -293,7 +293,7 @@ export class DatasetActions {
                 const newTemplate: Types.DataSetAllocTemplate = {
                     [templateName]: dsPropsForAPI,
                 };
-                datasetProvider.addDsTemplate(newTemplate);
+                await datasetProvider.addDsTemplate(newTemplate);
             }
         });
     }
@@ -1776,6 +1776,14 @@ export class DatasetActions {
                     await AuthUtils.errorHandling(error, DatasetUtils.getNodeLabels(node).dataSetName, vscode.l10n.t("Unable to copy data set."));
                 }
             }
+        }
+    }
+
+    public static async copyName(node: IZoweDatasetTreeNode): Promise<void> {
+        if (SharedContext.isDsMember(node) && node.getParent()) {
+            await vscode.env.clipboard.writeText(`${node.getParent().label as string}(${node.label as string})`);
+        } else if (SharedContext.isDs(node) || SharedContext.isPds(node) || SharedContext.isMigrated(node)) {
+            await vscode.env.clipboard.writeText(node.label as string);
         }
     }
 }