+++ packages/zowe-explorer/src/trees/uss/USSInit.ts
@@ -71,6 +71,7 @@ export class USSInit {
         );
         context.subscriptions.push(
             vscode.commands.registerCommand("zowe.uss.refreshUSS", async (node, nodeList) => {
+                const statusMsg = Gui.setStatusBarMessage(vscode.l10n.t("$(sync~spin) Pulling from Mainframe..."));
                 let selectedNodes = SharedUtils.getSelectedNodeList(node, nodeList) as IZoweUSSTreeNode[];
                 selectedNodes = selectedNodes.filter((x) => SharedContext.isDocument(x));
                 for (const item of selectedNodes) {
@@ -81,13 +82,14 @@ export class USSInit {
                         if (!(await FsAbstractUtils.confirmForUnsavedDoc(node.resourceUri))) {
                             return;
                         }
-                        const statusMsg = Gui.setStatusBarMessage("$(sync~spin) Fetching USS file...");
+                        const statusMsg2 = Gui.setStatusBarMessage("$(sync~spin) Fetching USS file...");
                         // need to pull content for file and apply to FS entry
                         await UssFSProvider.instance.fetchFileAtUri(item.resourceUri, {
                             editor: vscode.window.visibleTextEditors.find((v) => v.document.uri.path === item.resourceUri.path),
                         });
-                        statusMsg.dispose();
+                        statusMsg2.dispose();
                     }
+                    statusMsg.dispose();
                 }
             })
         );
@@ -215,6 +217,9 @@ export class USSInit {
                 "zowe.uss.openWithEncoding",
                 (node: IZoweUSSTreeNode, encoding?: ZosEncoding): Promise<void> => ussFileProvider.openWithEncoding(node, encoding)
             )
+        );
+        context.subscriptions.push(
+            vscode.commands.registerCommand("zowe.uss.copyRelativePath", async (node: IZoweUSSTreeNode) => USSActions.copyRelativePath(node))
         );
         context.subscriptions.push(
             vscode.workspace.onDidChangeConfiguration(async (e) => {